mpn_mod to its own module, continued
