+++ src/mpn_mod.h
@@ -21,6 +21,8 @@
 #include "flint.h"
 #include "mpn_extras.h"
 #include "gr.h"
+#include "gr_poly.h"
+#include "gr_mat.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -148,6 +150,186 @@ flint_mpn_submod_2(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_srcptr m)
     if (cmp)
         add_ssaaaa(res[1], res[0], res[1], res[0], m1, m0);
 }
+
+FLINT_FORCE_INLINE int
+flint_mpn_signed_sub_n(mp_ptr res, mp_srcptr x, mp_srcptr y, mp_size_t n)
+{
+    if (mpn_cmp(x, y, n) >= 0)
+    {
+        mpn_sub_n(res, x, y, n);
+        return 0;
+    }
+    else
+    {
+        mpn_sub_n(res, y, x, n);
+        return 1;
+    }
+}
+
+FLINT_FORCE_INLINE void
+flint_mpn_signed_div2(mp_ptr res, mp_srcptr x, mp_size_t n)
+{
+    mp_limb_t s = x[n - 1] & (UWORD(1) << (FLINT_BITS - 1));
+    mpn_rshift(res, x, n, 1);
+    res[n - 1] |= s;
+}
+
+void flint_mpn_mulmod_preinvn_2(mp_ptr r,
+        mp_srcptr a, mp_srcptr b,
+        mp_srcptr d, mp_srcptr dinv, ulong norm);
+
+char * _flint_mpn_get_str(mp_srcptr x, mp_size_t n);
+
+/* Basic operations and arithmetic */
+
+int gr_ctx_init_mpn_mod(gr_ctx_t ctx, const fmpz_t n);
+void gr_ctx_mpn_mod_set_primality(gr_ctx_t ctx, truth_t is_prime);
+
+int mpn_mod_ctx_write(gr_stream_t out, gr_ctx_t ctx);
+void mpn_mod_ctx_clear(gr_ctx_t ctx);
+
+MPN_MOD_INLINE truth_t
+mpn_mod_ctx_is_field(gr_ctx_t ctx)
+{
+    return MPN_MOD_CTX_IS_PRIME(ctx);
+}
+
+MPN_MOD_INLINE void
+mpn_mod_init(mp_ptr x, gr_ctx_t ctx)
+{
+    flint_mpn_zero(x, MPN_MOD_CTX_NLIMBS(ctx));
+}
+
+MPN_MOD_INLINE void
+mpn_mod_clear(mp_ptr x, gr_ctx_t ctx)
+{
+}
+
+MPN_MOD_INLINE void
+mpn_mod_swap(mp_ptr x, mp_ptr y, gr_ctx_t ctx)
+{
+    slong i = 0, n = MPN_MOD_CTX_NLIMBS(ctx);
+    for (i = 0; i < n; i++)
+        FLINT_SWAP(mp_limb_t, x[i], y[i]);
+}
+
+MPN_MOD_INLINE int
+mpn_mod_set(mp_ptr res, mp_srcptr x, gr_ctx_t ctx)
+{
+    flint_mpn_copyi(res, x, MPN_MOD_CTX_NLIMBS(ctx));
+    return GR_SUCCESS;
+}
+
+MPN_MOD_INLINE int
+mpn_mod_zero(mp_ptr res, gr_ctx_t ctx)
+{
+    flint_mpn_zero(res, MPN_MOD_CTX_NLIMBS(ctx));
+    return GR_SUCCESS;
+}
+
+MPN_MOD_INLINE
+int mpn_mod_one(mp_ptr res, gr_ctx_t ctx)
+{
+    res[0] = 1;
+    flint_mpn_zero(res + 1, MPN_MOD_CTX_NLIMBS(ctx) - 1);
+    return GR_SUCCESS;
+}
+
+int mpn_mod_set_ui(mp_ptr res, ulong x, gr_ctx_t ctx);
+int mpn_mod_set_si(mp_ptr res, slong x, gr_ctx_t ctx);
+int mpn_mod_neg_one(mp_ptr res, gr_ctx_t ctx);
+
+int mpn_mod_set_mpn(mp_ptr res, mp_srcptr x, mp_size_t xn, gr_ctx_t ctx);
+int mpn_mod_set_fmpz(mp_ptr res, const fmpz_t x, gr_ctx_t ctx);
+int mpn_mod_set_other(mp_ptr res, gr_ptr v, gr_ctx_t v_ctx, gr_ctx_t ctx);
+int mpn_mod_randtest(mp_ptr res, flint_rand_t state, gr_ctx_t ctx);
+int mpn_mod_write(gr_stream_t out, mp_srcptr x, gr_ctx_t ctx);
+
+int mpn_mod_get_fmpz(fmpz_t res, mp_srcptr x, gr_ctx_t ctx);
+
+MPN_MOD_INLINE truth_t
+mpn_mod_is_zero(mp_srcptr x, gr_ctx_t ctx)
+{
+    return flint_mpn_zero_p(x, MPN_MOD_CTX_NLIMBS(ctx)) ? T_TRUE : T_FALSE;
+}
+
+MPN_MOD_INLINE truth_t
+mpn_mod_is_one(mp_srcptr x, gr_ctx_t ctx)
+{
+    return (x[0] == 1 && flint_mpn_zero_p(x + 1, MPN_MOD_CTX_NLIMBS(ctx) - 1)) ? T_TRUE : T_FALSE;
+}
+
+truth_t mpn_mod_is_neg_one(gr_srcptr x, gr_ctx_t ctx);
+
+MPN_MOD_INLINE truth_t
+mpn_mod_equal(mp_srcptr x, mp_srcptr y, gr_ctx_t ctx)
+{
+    return flint_mpn_equal_p(x, y, MPN_MOD_CTX_NLIMBS(ctx)) ? T_TRUE : T_FALSE;
+}
+
+int mpn_mod_neg(mp_ptr res, mp_srcptr x, gr_ctx_t ctx);
+int mpn_mod_add(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+int mpn_mod_sub(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+int mpn_mod_add_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx);
+int mpn_mod_sub_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx);
+int mpn_mod_add_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx);
+int mpn_mod_sub_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx);
+int mpn_mod_add_fmpz(mp_ptr res, mp_srcptr x, const fmpz_t y, gr_ctx_t ctx);
+int mpn_mod_sub_fmpz(mp_ptr res, mp_srcptr x, const fmpz_t y, gr_ctx_t ctx);
+
+int mpn_mod_mul(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+
+int mpn_mod_mul_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx);
+int mpn_mod_mul_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx);
+int mpn_mod_mul_fmpz(mp_ptr res, mp_srcptr x, const fmpz_t y, gr_ctx_t ctx);
+int mpn_mod_addmul(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+int mpn_mod_addmul_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx);
+int mpn_mod_addmul_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx);
+int mpn_mod_addmul_fmpz(mp_ptr res, mp_srcptr x, const fmpz_t y, gr_ctx_t ctx);
+int mpn_mod_submul(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+int mpn_mod_submul_ui(mp_ptr res, mp_srcptr x, ulong y, gr_ctx_t ctx);
+int mpn_mod_submul_si(mp_ptr res, mp_srcptr x, slong y, gr_ctx_t ctx);
+int mpn_mod_submul_fmpz(mp_ptr res, mp_srcptr x, const fmpz_t y, gr_ctx_t ctx);
+
+MPN_MOD_INLINE int
+mpn_mod_sqr(mp_ptr res, mp_srcptr x, gr_ctx_t ctx)
+{
+    return mpn_mod_mul(res, x, x, ctx);
+}
+
+int mpn_mod_inv(mp_ptr res, mp_srcptr x, gr_ctx_t ctx);
+int mpn_mod_div(mp_ptr res, mp_srcptr x, mp_srcptr y, gr_ctx_t ctx);
+
+/* Vector functions */
+
+int _mpn_mod_vec_zero(mp_ptr res, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_clear(mp_ptr res, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_set(mp_ptr res, mp_srcptr x, slong len, gr_ctx_t ctx);
+void _mpn_mod_vec_swap(mp_ptr vec1, mp_ptr vec2, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_neg(mp_ptr res, mp_srcptr x, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_add(mp_ptr res, mp_srcptr x, mp_srcptr y, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_sub(mp_ptr res, mp_srcptr x, mp_srcptr y, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_mul(mp_ptr res, mp_srcptr x, mp_srcptr y, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_mul_scalar(mp_ptr res, mp_srcptr x, slong len, mp_srcptr y, gr_ctx_t ctx);
+int _mpn_mod_scalar_mul_vec(mp_ptr res, mp_srcptr y, mp_srcptr x, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_addmul_scalar(mp_ptr res, mp_srcptr x, slong len, mp_srcptr y, gr_ctx_t ctx);
+int _mpn_mod_vec_dot(mp_ptr res, mp_srcptr initial, int subtract, mp_srcptr vec1, mp_srcptr vec2, slong len, gr_ctx_t ctx);
+int _mpn_mod_vec_dot_rev(mp_ptr res, mp_srcptr initial, int subtract, mp_srcptr vec1, mp_srcptr vec2, slong len, gr_ctx_t ctx);
+
+/* Matrix algorithms */
+
+int mpn_mod_mat_mul_waksman(gr_mat_t C, const gr_mat_t A, const gr_mat_t B, gr_ctx_t ctx);
+int mpn_mod_mat_mul_multi_mod(gr_mat_t C, const gr_mat_t A, const gr_mat_t B, gr_ctx_t ctx);
+int mpn_mod_mat_mul(gr_mat_t C, const gr_mat_t A, const gr_mat_t B, gr_ctx_t ctx);
+
+int mpn_mod_mat_nonsingular_solve_tril(gr_mat_t X, const gr_mat_t L, const gr_mat_t B, int unit, gr_ctx_t ctx);
+int mpn_mod_mat_nonsingular_solve_triu(gr_mat_t X, const gr_mat_t U, const gr_mat_t B, int unit, gr_ctx_t ctx);
+int mpn_mod_mat_lu_classical_delayed(slong * res_rank, slong * P, gr_mat_t A, const gr_mat_t A_in, int rank_check, gr_ctx_t ctx);
+int mpn_mod_mat_lu(slong * rank, slong * P, gr_mat_t LU, const gr_mat_t A, int rank_check, gr_ctx_t ctx);
+int mpn_mod_mat_det(mp_ptr res, const gr_mat_t A, gr_ctx_t ctx);
+
+/* Polynomial algorithms */
+/* TODO */
 
 #ifdef __cplusplus
 }
